scalekit_agent_connect/requirements.txt
a_beta_scalekit==3.0.1
langchain_core==0.3.74
langchain_mcp_adapters==0.1.9
langchain_openai==0.3.30
langgraph==0.6.5
